machinelearning/konlp_transfomer/requirements.txt
pandas
numpy
scikit-learn
scipy
tqdm
implicit
threadpoolctl
sqlalchemy
pyarrow # (pickle 데이터 대신 parquet/csv를 다룰 수도 있음)
torch
openpyxl
matplotlib
sentence-transformers
seaborn
umap-learn